Object Object

Object Object

Object(acl = NULL, bucket = NULL, cacheControl = NULL,
  componentCount = NULL, contentDisposition = NULL,
  contentEncoding = NULL, contentLanguage = NULL, contentType = NULL,
  crc32c = NULL, customerEncryption = NULL, etag = NULL,
  generation = NULL, id = NULL, md5Hash = NULL, mediaLink = NULL,
  metadata = NULL, metageneration = NULL, name = NULL, owner = NULL,
  selfLink = NULL, size = NULL, storageClass = NULL, timeCreated = NULL,
  timeDeleted = NULL, updated = NULL)

Arguments

acl
Access controls on the object

bucket
The name of the bucket containing this object

cacheControl
Cache-Control directive for the object data

componentCount
Number of underlying components that make up this object

contentDisposition
Content-Disposition of the object data

contentEncoding
Content-Encoding of the object data

contentLanguage
Content-Language of the object data

contentType
Content-Type of the object data

crc32c
CRC32c checksum, as described in RFC 4960, Appendix B; encoded using base64 in big-endian byte order

customerEncryption
Metadata of customer-supplied encryption key, if the object is encrypted by such a key

etag
HTTP 1

generation
The content generation of this object

id
The ID of the object

md5Hash
MD5 hash of the data; encoded using base64

mediaLink
Media download link

metadata
User-provided metadata, in key/value pairs

metageneration
The version of the metadata for this object at this generation

name
The name of this object

owner
The owner of the object

selfLink
The link to this object

size
Content-Length of the data in bytes

storageClass
Storage class of the object

timeCreated
The creation time of the object in RFC 3339 format

timeDeleted
The deletion time of the object in RFC 3339 format

updated
The modification time of the object metadata in RFC 3339 format

Value

Object object

Details

An object.