Create a Bucket Access Controls

Create a new access control at the bucket level

gcs_create_bucket_acl(
  bucket = gcs_get_global_bucket(),
  entity = "",
  entity_type = c("user", "group", "domain", "project", "allUsers",
    "allAuthenticatedUsers"),
  role = c("READER", "OWNER")
)

Arguments

bucket
Name of a bucket, or a bucket object returned by gcs_create_bucket

entity
The entity holding the permission. Not needed for entity_type allUsers or allAuthenticatedUsers

entity_type
what type of entity

role
Access permission for entity

Used also for when a bucket is updated

Value

Bucket access control object

See also