Create a lifecycle condition

Use this to set rules for how long objects last in a bucket in gcs_create_bucket

gcs_create_lifecycle(age = NULL, createdBefore = NULL,
  numNewerVersions = NULL, isLive = NULL)

Arguments

age
Age in days before objects are deleted

createdBefore
Deletes all objects before this date

numNewerVersions
Deletes all newer versions of this object

isLive
If TRUE deletes all live objects, if FALSE deletes all archived versions numNewerVersions and isLive works only for buckets with object versioning

For multiple conditions, pass this object in as a list.

See also

Lifecycle documentation https://cloud.google.com/storage/docs/lifecycle

Other bucket functions: gcs_create_bucket, gcs_delete_bucket, gcs_get_bucket, gcs_get_global_bucket, gcs_global_bucket, gcs_list_buckets