Get the download URL

Create the download URL for objects in buckets

gcs_download_url(object_name, bucket = gcs_get_global_bucket(),
  public = FALSE)

Arguments

object_name
A vector of object names

bucket
A vector of bucket names

public
TRUE to return a public URL

Value

the URL for downloading objects

Details

bucket names should be length 1 or same length as object_name

Download URLs can be either authenticated behind a login that you may need to update access for via gcs_update_object_acl, or public to all if their predefinedAcl = 'publicRead'

Use the public = TRUE to return the URL accessible to all, which changes the domain name from storage.cloud.google.com to storage.googleapis.com

See also

Other download functions: gcs_parse_download, gcs_signed_url