List objects in a bucket

List objects in a bucket

gcs_list_objects(
  bucket = gcs_get_global_bucket(),
  detail = c("summary", "more", "full"),
  prefix = NULL,
  delimiter = NULL
)

Arguments

bucket
bucket containing the objects

detail
Set level of detail

prefix
Filter results to objects whose names begin with this prefix

delimiter
Use to list objects like a directory listing.

Value

A data.frame of the objects

Details

Columns returned by detail are:

summary - name, size, updated

more - as above plus: bucket, contentType, storageClass, timeCreated

full - as above plus: id, selfLink, generation, metageneration, md5Hash, mediaLink, crc32c, etag

delimited returns results in a directory-like mode: items will contain only objects whose names, aside from the prefix, do not contain delimiter. In conjunction with the prefix filter, the use of the delimiter parameter allows the list method to operate like a directory listing, despite the object namespace being flat. For example, if delimiter were set to "/", then listing objects from a bucket that contains the objects "a/b", "a/c", "dddd", "eeee", "e/f" would return objects "dddd" and "eeee", and prefixes "a/" and "e/".

See also