List pub/sub notifications for a bucket

List notification configurations for a bucket.

gcs_list_pubsub(bucket = gcs_get_global_bucket())

Arguments

bucket
The bucket for notifications

Details

Cloud Pub/Sub notifications allow you to track changes to your Cloud Storage objects. As a minimum you wil need: the Cloud Pub/Sub API activated for the project; sufficient permissions on the bucket you wish to monitor; sufficient permissions on the project to receive notifications; an existing pub/sub topic; have given your service account at least pubsub.publisher permission.

See also