Load .RData objects or sessions from the Google Cloud

Load R objects that have been saved using gcs_save or gcs_save_image

gcs_load(file = ".RData", bucket = gcs_get_global_bucket(),
  envir = .GlobalEnv, saveToDisk = file, overwrite = TRUE)

Arguments

file
Where the files are stored

bucket
Bucket the stored objects are in

envir
Environment to load objects into

saveToDisk
Where to save the loaded file. Default same file name

overwrite
If file exists, overwrite. Default TRUE.

Value

TRUE if successful

Details

The argument file's default is to load an image file called .RData from gcs_save_image into the Global environment.

This would overwrite your existing .RData file in the working directory, so change the file name if you don't wish this to be the case.

See also

Other R session data functions: gcs_save_all, gcs_save_image, gcs_save, gcs_source