Retry a resumeable upload

Used internally in gcs_upload, you can also use this for failed uploads within one week of generating the upload URL

gcs_retry_upload(retry_object = NULL, upload_url = NULL, file = NULL,
  type = NULL)

Arguments

retry_object
A object of class gcs_upload_retry.

upload_url
As created in a failed upload via gcs_upload

file
The file location to upload

type
The file type, guessed if NULL

Either supply a retry object, or the upload_url, file and type manually yourself.

The function will first check to see how much has been uploaded already, then try to send up the remaining bytes.

Value

If successful, an object metadata object, if not an gcs_upload_retry object.