Save .RData objects to the Google Cloud

Performs save then saves it to Google Cloud Storage.

gcs_save(..., file, bucket = gcs_get_global_bucket(),
  envir = parent.frame())

Arguments

...
The names of the objects to be saved (as symbols or character strings).

file
The file name that will be uploaded (conventionally with file extension .RData)

bucket
Bucket to store objects in

envir
Environment to search for objects to be saved

Value

The GCS object

Details

For all session data use gcs_save_image instead. gcs_save(ob1, ob2, ob3, file = "mydata.RData") will save the objects specified to an .RData file then save it to Cloud Storage, to be loaded later using gcs_load.

For any other use, its better to use gcs_upload and gcs_get_object instead.

Restore the R objects using gcs_load(bucket = "your_bucket")

This will overwrite any data within your local environment with the same name.

See also

Other R session data functions: gcs_load, gcs_save_all, gcs_save_image, gcs_source