Save an R session to the Google Cloud

Performs save.image then saves it to Google Cloud Storage.

gcs_save_image(file = ".RData", bucket = gcs_get_global_bucket(),
  saveLocation = NULL, envir = parent.frame())

Arguments

file
Where to save the file in GCS and locally

bucket
Bucket to store objects in

saveLocation
Which folder in the bucket to save file

envir
Environment to save from

Value

The GCS object

Details

gcs_save_image(bucket = "your_bucket") will save all objects in the workspace to .RData folder on Google Cloud Storage within your_bucket.

Restore the objects using gcs_load(bucket = "your_bucket")

This will overwrite any data with the same name in your current local environment.

See also

Other R session data functions: gcs_load, gcs_save_all, gcs_save, gcs_source