Change access to an object in a bucket

Updates Google Cloud Storage ObjectAccessControls

gcs_update_object_acl(object_name, bucket = gcs_get_global_bucket(),
  entity = "", entity_type = c("user", "group", "domain", "project",
  "allUsers", "allAuthenticatedUsers"), role = c("READER", "OWNER"))

Arguments

object_name
Object to update

bucket
Google Cloud Storage bucket

entity
entity to update or add, such as an email

entity_type
what type of entity

role
Access permission for entity

Value

TRUE if successful

Details

An entity is an identifier for the entity_type.

entity="user" may have userId or email

entity="group" may have groupId or email

entity="domain" may have domain

entity="project" may have team-projectId

For example:

entity="user" could be jane@doe.com

entity="group" could be example@googlegroups.com

entity="domain" could be example.com which is a Google Apps for Business domain.

See also